Added functions and a prompt
The first function added was getComputerChoice.
It is used to randomly choose a number and return
an element from the array with rock, paper, and
scissor at index ranNum.

Second function is playRound and this is where
the playerSelection and computerSelection is
used. The to inputs are used to find who lost,
who won, or was it a tie in the round of rock,
paper, scissor.

Third function capitalize is used to make the user
input uniform with computerSelection. Uppercase
first letter and the rest are lowercase.

Prompt was added for input from the user.